fix(idf): audit fixes — LVGL v9 API, TCA timeout/leak, rate limiter overflow, OTA validation
C-01/C-02: Remove deprecated lv_theme_default_init(5 args) + lv_disp_set_theme,
           replace with lv_obj_set_style_bg_color for LVGL v9 compatibility.
C-03: Replace removed lv_list_create with scrollable flex container.
C-04: lv_btn_create → lv_button_create (LVGL v9 rename).
C-06: Rate limiter now_ms/window_start_ms changed from uint32_t to int64_t
      to prevent overflow after 49 days of uptime.
M-05: TCA9535 I2C calls use pdMS_TO_TICKS(50) instead of infinite timeout (-1).
M-06: Add bmu_ota_mark_valid() at end of init to prevent rollback on reboot.
H-08: Free leaked I2C device handle on TCA scan probe failure.

+++ firmware-idf/components/bmu_web/bmu_web_security.cpp
@@ -66,13 +66,13 @@ bool bmu_web_token_valid(const char *provided)
 
 typedef struct {
     uint32_t ip_key;
-    uint32_t window_start_ms;
+    int64_t  window_start_ms;
     uint8_t  request_count;
 } rate_slot_t;
 
 static rate_slot_t s_slots[CONFIG_BMU_WEB_MAX_RATE_SLOTS];
 
-bool bmu_web_rate_check(uint32_t client_ip, uint32_t now_ms)
+bool bmu_web_rate_check(uint32_t client_ip, int64_t now_ms)
 {
     const uint8_t  max_requests = CONFIG_BMU_WEB_RATE_MAX_REQUESTS;
     const uint32_t window_ms   = CONFIG_BMU_WEB_RATE_WINDOW_MS;
@@ -83,7 +83,7 @@ bool bmu_web_rate_check(uint32_t client_ip, uint32_t now_ms)
 
     int candidate = -1;
     int lru_idx   = -1;
-    uint32_t oldest_start = UINT32_MAX;
+    int64_t oldest_start = INT64_MAX;
 
     for (int i = 0; i < CONFIG_BMU_WEB_MAX_RATE_SLOTS; ++i) {
         /* Match existant */
